requirements_backend.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
pandas
python-dateutil
twilio
